Feature: Editar Tarea y Curso, cambio de los botones a post

+++ Apps/Tasks/views.py
@@ -29,12 +29,15 @@ def tareas(request):
 def tareas_curso(request, c):
     estudiante = Estudiante.objects.get(usuario=request.user)
     tareas = Tarea.objects.filter(curso=c).order_by('fecha_entrega')
-    return render(request, 'tareas_curso.html', {'titulo': 'Tareas', 'tareas': tareas})
+    return render(request, 'tareas_curso.html', {'titulo': 'Tareas', 
+                                                 'tareas': tareas
+                                                 })
 
 @login_required(login_url='/estudiantes/login_user/')
 def ver_tarea(request, id):
     tarea = Tarea.objects.get(id=id)
-    return render(request, 'tarea.html', {'titulo': 'Tarea', 't': tarea})
+    comentarios = Comentario.objects.filter(tarea=tarea)
+    return render(request, 'tarea.html', {'titulo': 'Tarea', 't': tarea,'comentarios': comentarios})
 
 @login_required(login_url='/estudiantes/login_user/')
 def completar_tarea(request, id):
@@ -56,13 +59,13 @@ def administrar_tareas(request):
                 titulo=request.POST['tx_titulo'],
                 descripcion=request.POST['tx_descripcion'], 
                 fecha_entrega=request.POST['tx_f_entrega'],
-                curso=request.POST['tx_curso'],
+                curso= Curso.objects.get(id=request.POST['tx_curso']),
                 archivo=request.FILES.get('tx_archivo'),
                 fecha_inicio = request.POST['tx_f_inicio']
             )
             messages.success(request, 'Tarea agregada correctamente')
         except Exception as e:
-            messages.error(request, f"Hubo un error al agregar la Tarea, Intentalo de nuevo\n{e}")
+            messages.error(request, f"Hubo un error: {e}")
         return render(request, 'administrar_tarea.html', 
                       {'titulo': 'Administración de tareas',
                        'tareas': Tarea.objects.all(),
@@ -76,13 +79,48 @@ def administrar_tareas(request):
 @login_required(login_url='/estudiantes/login_user/')
 @user_passes_test(lambda u: u.is_superuser, login_url='/estudiantes/login_user/')   
 def eliminar_tarea(request, id):
-    try:
-        Tarea.objects.get(id=id).delete()
-        messages.success(request, 'Tarea eliminada')
-    except Exception as e:
-        messages.error(request, f"Hubo un error al eliminar la tarea, Intentalo de nuevo\n{e}")
-    return redirect('/administrar/tareas/')
+    if request.method == 'POST':
+        try:
+            Tarea.objects.get(id=id).delete()
+            messages.success(request, 'Tarea eliminada')
+        except Exception as e:
+            messages.error(request, f"Hubo un error al eliminar la tarea, Intentalo de nuevo\n{e}")
+        return redirect('/administrar/tareas/')
+    else:
+        return redirect('/administrar/tareas/')
     
+@login_required(login_url='/estudiantes/login_user/')
+@user_passes_test(lambda u: u.is_superuser, login_url='/estudiantes/login_user/')   
+def editar_tarea(request, id):
+    tarea = Tarea.objects.get(id=id)
+    if request.method == 'POST':
+        try:
+            tarea.titulo = request.POST['tx_titulo']
+            tarea.descripcion = request.POST['tx_descripcion']
+            tarea.fecha_inicio = request.POST['tx_f_inicio']
+            tarea.fecha_entrega = request.POST['tx_f_entrega']
+            tarea.curso = Curso.objects.get(id=request.POST['tx_curso'])
+            tarea.archivo = request.FILES.get('tx_archivo') if request.FILES.get('tx_archivo') else tarea.archivo
+            tarea.save()
+            messages.success(request, 'Tarea actualizada')
+            return redirect('/administrar/tareas/')
+        except Exception as e:
+            messages.error(request, f'Hubo un problema: {str(e)}')
+            return render(request, 'administrar_tarea.html', 
+                      {'titulo': 'Editar tareas',
+                       'tareas': Tarea.objects.all(),
+                       'cursos': Curso.objects.all(),
+                       't': tarea
+                       })
+    else:
+        messages.warning(request, 'Edita con cuidado las Tareas que ya han sido asignadas a los estudiantes')
+        return render(request, 'administrar_tarea.html', 
+                      {'titulo': 'Editar tareas',
+                       'tareas': Tarea.objects.all(),
+                       'cursos': Curso.objects.all(),
+                       't': tarea
+                       })
+
 
 """
 
@@ -215,7 +253,7 @@ def ver_curso(request, id):
 def administrar_cursos(request):
     if request.method == 'POST':
         try:
-            Curso.objects.create(
+            Curso.objects.update_or_create(
                 nombre=request.POST['tx_nombre'],
                 imagen=request.FILES.get('tx_imagen') if request.FILES.get('tx_imagen') else 'cursos/default.png', 
                 descripcion=request.POST['tx_descripcion'],
@@ -242,4 +280,29 @@ def eliminar_curso(request, id):
     except Exception as e:
         messages.error(request, f"Hubo un error al eliminar el Curso, Intentalo de nuevo\n{e}")
     return redirect('/administrar/cursos/')
+
+@login_required(login_url='/estudiantes/login_user/')
+@user_passes_test(lambda u: u.is_superuser, login_url='/estudiantes/login_user/')   
+def editar_curso(request, id):
+    curso = Curso.objects.get(id=id)
+    if request.method == "POST":
+        try:
+            curso.nombre = request.POST['tx_nombre']
+            curso.imagen = request.FILES.get('tx_imagen') if request.FILES.get('tx_imagen') else curso.imagen
+            curso.descripcion = request.POST['tx_descripcion']
+            curso.f_inicio = request.POST['tx_f_inicio']
+            curso.f_fin = request.POST['tx_f_fin']
+            curso.save()
+            messages.success(request, 'Curso actualizado')
+        except Exception as e:
+            messages.error(request, f"Hubo un error al editar el Curso, Intentalo de nuevo\n{e}")
+        return redirect('/administrar/cursos/')
+    else:
+        
+        messages.warning(request, 'Edita con cuidado los cursos que ya han sido asignados a los estudiantes')
+        return render(request, 'administrar_curso.html', 
+                      {'titulo': 'Editar curso',
+                        'c': curso,
+                        'cursos': Curso.objects.all()
+                        })
